+++ lib/actions/authorization/process_response_types.js
@@ -10,14 +10,17 @@ module.exports = (provider) => {
   } = instance(provider).configuration();
 
   async function tokenHandler(ctx) {
+    const accountId = ctx.oidc.session.accountId();
     const at = new AccessToken({
-      accountId: ctx.oidc.session.accountId(),
+      accountId,
       claims: ctx.oidc.claims,
       clientId: ctx.oidc.client.clientId,
       grantId: ctx.oidc.uuid,
       scope: ctx.oidc.params.scope,
       sid: ctx.oidc.session.sidFor(ctx.oidc.client.clientId),
     });
+
+    at.setAudiences(await audiences(ctx, accountId, at, 'access_token', ctx.oidc.params.scope));
 
     ctx.oidc.entity('AccessToken', at);
 
@@ -112,7 +115,7 @@ module.exports = (provider) => {
 
     if (res.id_token) {
       res.id_token = await res.id_token.sign(ctx.oidc.client, {
-        audiences: await audiences(ctx, ctx.oidc.account.accountId, 'id_token', ctx.oidc.params.scope),
+        audiences: await audiences(ctx, ctx.oidc.session.accountId(), undefined, 'id_token', ctx.oidc.params.scope),
       });
     }
