+++ lib/actions/grants/authorization_code.js
@@ -77,6 +77,8 @@ module.exports.handler = function getAuthorizationCodeHandler(provider) {
       sid: code.sid,
     });
 
+    at.setAudiences(await audiences(ctx, account.accountId, at, 'access_token', code.scope));
+
     const accessToken = await at.save();
     ctx.oidc.entity('AccessToken', at);
