+++ lib/actions/grants/refresh_token.js
@@ -80,14 +80,17 @@ module.exports.handler = function getRefreshTokenHandler(provider) {
       }
     }
 
+    const scope = ctx.oidc.params.scope || refreshToken.scope;
     const at = new AccessToken({
+      scope,
       accountId: account.accountId,
       claims: refreshToken.claims,
       clientId: ctx.oidc.client.clientId,
       grantId: refreshToken.grantId,
-      scope: ctx.oidc.params.scope || refreshToken.scope,
       sid: refreshToken.sid,
     });
+
+    at.setAudiences(await audiences(ctx, account.accountId, at, 'access_token', scope));
 
     const accessToken = await at.save();
     ctx.oidc.entity('AccessToken', at);
