+++ lib/actions/introspection.js
@@ -134,6 +134,7 @@ module.exports = function introspectionAction(provider) {
         sid: token.sid,
         iss: token.iss,
         jti: token.jti,
+        aud: token.aud,
         scope: token.scope,
       });
