+++ lib/helpers/defaults.js
@@ -509,8 +509,8 @@ const DEFAULTS = {
     // @param sub   - account identifier (subject)
     // @param token - a reference to the token used for which a given account is being loaded,
     //   is undefined in scenarios where claims are returned from authorization endpoint
-    // @param use   - can either be "id_token" or "userinfo", depending on
-    //   where the specific audiences are intended to be put in
+    // @param use   - can be one of "id_token", "userinfo", "access_token" depending on where the
+    //   specific audiences are intended to be put in
     // @param scope - scope from either the request or related token
     return undefined;
   },
