+++ lib/models/base_token.js
@@ -9,6 +9,7 @@ const IN_PAYLOAD = [
   'codeChallengeMethod', // for authorization code
   'grantId',
   'jti',
+  'aud',
   'kind',
   'nonce',
   'redirectUri',
