+++ lib/models/id_token.js
@@ -7,19 +7,9 @@ const epochTime = require('../helpers/epoch_time');
 const JWT = require('../helpers/jwt');
 const { InvalidClientMetadata } = require('../helpers/errors');
 const instance = require('../helpers/weak_cache');
+const ensureConform = require('../helpers/ensure_conform');
 
 const hashes = ['at_hash', 'c_hash', 'rt_hash', 's_hash'];
-
-function ensureConform(audiences, clientId) {
-  assert(Array.isArray(audiences), 'audiences must be an array');
-
-  const value = Array.from(audiences);
-  if (!value.includes(clientId)) {
-    value.unshift(clientId);
-  }
-
-  return value;
-}
 
 module.exports = function getIdToken(provider) {
   const Claims = getMask(instance(provider).configuration());
